pyproject.toml
[project]
name = "reservation-notifier"
version = "0.1.0"
description = ""
authors = [
    {name = "GiuliaZobrist",email = "giulia-zobrist@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "requests-mock (>=1.12.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "reservation_notifier", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
